refactor name of module containing dynamodb connector
